requirements.txt
beautifulsoup4==4.4.1
Django==1.8.4
django-autofixture==0.10.1
django-haystack==2.4.0
docutils==0.12
Jinja2==2.8
lxml==3.4.4
MarkupSafe==0.23
MySQL-python==1.2.5
pockets==0.2.4
Pygments==2.0.2
pysolr==3.3.2
python-docx==0.8.5
requests==2.7.0
six==1.9.0
wheel==0.24.0
